Add `--run-docker` flag to execute goad in docker containers
When executing goad with the `--run-docker` flag it will try to connect
to the local docker daemon and pull the required image.

Instead of launching the default runner in a series of lambda functions
on aws, goad will execute the goad runner inside a docker image on the
locally configured docker daemon.

It will the folder beeing executed right now is the `data/lambda` folder
relative to the current working directory to work seamlessly with the
current development version of goad. When exending the ini
sepecification well make this more flexible to allow different runners
to be configured with goad.

The new aws sessions will be created for each docker container and
passed into it to allow connecting to the sqs queue launched on aws so
you can receive the results of the individual containers running on your
local machine.

Right now the inital run prints out the raw response of the docker api
beeing run while pulling the docker image. This is not looking great
yet, but well have to touch the current cli terminal library beeing
used to make this look nicer.