Use rabbit message queue for running on docker
With this commit launching goad with the run-docker flag will launch
rabbitmq in docker a local container with a seperate bridge network and
connect the lambdas running inside their containers to it.

This allows for execution without aws an account.

After running the infrasturcture set up will be teared down again.