Rewrite metrics calculation and retrieval
This commit moves most of the aggregation logic into the cli leaving the
goad runners with a cleaner implementation of the metrics to transmit to
the cli.

We introduce a runner id that allows tracking of the individual
lambdas being executed on aws and later aggregation of results for
different aws regions.

The infrastructure package has been restructured to move more of the aws
specific functionality into the corresponding packages.

A VSCode debug configuration has been added to version control.

Closes #143